refactor for easier code reuse
