+++ cmd/kubeless/function/top.go
@@ -17,26 +17,21 @@ limitations under the License.
 package function
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
 	"sort"
-	"strconv"
 	"time"
 
 	"github.com/ghodss/yaml"
 	"github.com/gosuri/uitable"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 
 	kubelessApi "github.com/kubeless/kubeless/pkg/apis/kubeless/v1beta1"
 	"github.com/kubeless/kubeless/pkg/client/clientset/versioned"
 	"github.com/kubeless/kubeless/pkg/utils"
-
-	"github.com/prometheus/common/expfmt"
 )
 
 var topCmd = &cobra.Command{
@@ -63,7 +58,7 @@ var topCmd = &cobra.Command{
 
 		apiV1Client := utils.GetClientOutOfCluster()
 		kubelessClient, err := utils.GetKubelessClientOutCluster()
-		handler := &PrometheusMetricsHandler{}
+		handler := &utils.PrometheusMetricsHandler{}
 
 		err = doTop(cmd.OutOrStdout(), kubelessClient, apiV1Client, handler, ns, functionName, output)
 		if err != nil {
@@ -72,152 +67,26 @@ var topCmd = &cobra.Command{
 	},
 }
 
-// Metric contains metrics for a functions
-type Metric struct {
-	FunctionName         string  `json:"function,omitempty"`
-	Namespace            string  `json:"namespace,omitempty"`
-	Method               string  `json:"method,omitempty"`
-	TotalCalls           float64 `json:"total_calls,omitempty"`
-	TotalFailures        float64 `json:"total_failures,omitempty"`
-	TotalDurationSeconds float64 `json:"total_duration_seconds,omitempty"`
-	AvgDurationSeconds   float64 `json:"avg_duration_seconds,omitempty"`
-}
-
 func init() {
 	topCmd.Flags().StringP("namespace", "n", "", "Specify namespace for the function")
 	topCmd.Flags().StringP("function", "f", "", "Specify the function")
 	topCmd.Flags().StringP("out", "o", "", "Output format. One of: json|yaml")
 }
 
-// MetricsRetriever is an interface for retreiving metrics from an endpoint
-type MetricsRetriever interface {
-	getRawMetrics(kubernetes.Interface, string, string) ([]byte, error)
-}
-
-// PrometheusMetricsHandler is a handler for retreiving metrics from Prometheus
-type PrometheusMetricsHandler struct{}
-
-func getFunctions(kubelessClient versioned.Interface, namespace, functionName string) ([]*kubelessApi.Function, error) {
-	var functionList *kubelessApi.FunctionList
-	var f *kubelessApi.Function
-	var err error
-	if functionName == "" {
-		functionList, err = kubelessClient.KubelessV1beta1().Functions(namespace).List(metav1.ListOptions{})
-		if err != nil {
-			return []*kubelessApi.Function{}, err
-		}
-	} else {
-		f, err = kubelessClient.KubelessV1beta1().Functions(namespace).Get(functionName, metav1.GetOptions{})
-		if err != nil {
-			return []*kubelessApi.Function{}, err
-		}
-		functionList = &kubelessApi.FunctionList{
-			Items: []*kubelessApi.Function{
-				f,
-			},
-		}
-	}
-	return functionList.Items, nil
-}
-
-func parseMetrics(namespace, functionName string, rawMetrics []byte) ([]*Metric, error) {
-	parser := expfmt.TextParser{}
-	parsedData, err := parser.TextToMetricFamilies(bytes.NewReader(rawMetrics))
-	if err != nil {
-		return nil, err
-	}
-
-	tmp := map[string]*Metric{}
-	var parsedMetrics []*Metric
-
-	metricsOfInterest := []string{"function_duration_seconds", "function_calls_total", "function_failures_total"}
-	for _, m := range metricsOfInterest {
-		for _, metric := range parsedData[m].GetMetric() {
-			// a function can have metrics for multiple methods (GET, POST, etc.)
-			// method names can be values other than GET/POST/PUT/DELETE
-			for _, label := range metric.GetLabel() {
-				if label.GetName() == "method" {
-					if _, ok := tmp[label.GetValue()]; !ok {
-						tmp[label.GetValue()] = &Metric{
-							FunctionName: functionName,
-							Namespace:    namespace,
-							Method:       label.GetValue(),
-						}
-					}
-					if m == "function_failures_total" {
-						tmp[label.GetValue()].TotalFailures = metric.GetCounter().GetValue()
-					}
-					if m == "function_duration_seconds" {
-						tmp[label.GetValue()].TotalDurationSeconds = metric.GetHistogram().GetSampleSum()
-					}
-					if m == "function_calls_total" {
-						tmp[label.GetValue()].TotalCalls = metric.GetCounter().GetValue()
-						if tmp[label.GetValue()].TotalCalls > 0 {
-							tmp[label.GetValue()].AvgDurationSeconds = float64(tmp[label.GetValue()].TotalDurationSeconds) / tmp[label.GetValue()].TotalCalls
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// if the funciton hasn't been invoked, add an item to the list so the function displays in the output
-	if len(tmp) == 0 {
-		tmp[""] = &Metric{
-			FunctionName: functionName,
-			Namespace:    namespace,
-		}
-	}
-
-	for _, v := range tmp {
-		parsedMetrics = append(parsedMetrics, v)
-	}
-
-	return parsedMetrics, nil
-}
-
-func (h *PrometheusMetricsHandler) getRawMetrics(apiV1Client kubernetes.Interface, namespace, functionName string) ([]byte, error) {
-	svc, err := apiV1Client.CoreV1().Services(namespace).Get(functionName, metav1.GetOptions{})
-	if err != nil {
-		return []byte{}, err
-	}
-
-	port := strconv.Itoa(int(svc.Spec.Ports[0].Port))
-	if svc.Spec.Ports[0].Name != "" {
-		port = svc.Spec.Ports[0].Name
-	}
-	req := apiV1Client.CoreV1().RESTClient().Get().Namespace(namespace).Resource("services").SubResource("proxy").Name(functionName + ":" + port).Suffix("/metrics")
-	return req.Do().Raw()
-}
-
-func getFunctionMetrics(apiV1Client kubernetes.Interface, h MetricsRetriever, namespace, functionName string) []*Metric {
-
-	res, err := h.getRawMetrics(apiV1Client, namespace, functionName)
-	if err != nil {
-		return []*Metric{}
-	}
-
-	metrics, err := parseMetrics(namespace, functionName, res)
-	if err != nil {
-		return []*Metric{}
-	}
-	return metrics
-}
-
-func doTop(w io.Writer, kubelessClient versioned.Interface, apiV1Client kubernetes.Interface, handler MetricsRetriever, ns, functionName, output string) error {
+func doTop(w io.Writer, kubelessClient versioned.Interface, apiV1Client kubernetes.Interface, handler utils.MetricsRetriever, ns, functionName, output string) error {
 	functions, err := getFunctions(kubelessClient, ns, functionName)
 	if err != nil {
 		return fmt.Errorf("Error listing functions: %v", err)
 	}
 
-	ch := make(chan []*Metric, len(functions))
+	ch := make(chan []*utils.Metric, len(functions))
 	for _, f := range functions {
 		go func(f *kubelessApi.Function) {
-			ch <- getFunctionMetrics(apiV1Client, handler, ns, f.ObjectMeta.Name)
+			ch <- utils.GetFunctionMetrics(apiV1Client, handler, ns, f.ObjectMeta.Name)
 		}(f)
 	}
 
-	var metrics []*Metric
+	var metrics []*utils.Metric
 
 	i := 0
 	for i < len(functions) {
@@ -238,7 +107,7 @@ func doTop(w io.Writer, kubelessClient versioned.Interface, apiV1Client kubernet
 	return printTop(w, metrics, apiV1Client, output)
 }
 
-func printTop(w io.Writer, metrics []*Metric, cli kubernetes.Interface, output string) error {
+func printTop(w io.Writer, metrics []*utils.Metric, cli kubernetes.Interface, output string) error {
 	if output == "" {
 		table := uitable.New()
 		table.MaxColWidth = 50
