+++ examples/Makefile
@@ -3,6 +3,7 @@ get-python:
 
 get-python-verify:
 	kubeless function call get-python |egrep hello.world
+	kubeless function top --function get-python | egrep function_calls_total.*1
 
 get-python-update:
 	$(eval TMPDIR := $(shell mktemp -d))
