+++ pkg/utils/event_sender.go
@@ -35,13 +35,23 @@ func IsJSON(s string) bool {
 
 }
 
+// GetFunctionPort returns the port for a function service
+func GetFunctionPort(clientset kubernetes.Interface, namespace, functionName string) (string, error) {
+	svc, err := clientset.CoreV1().Services(namespace).Get(functionName, metav1.GetOptions{})
+	if err != nil {
+		return "", fmt.Errorf("Unable to find the service for function %s", functionName)
+	}
+	return strconv.Itoa(int(svc.Spec.Ports[0].Port)), nil
+}
+
 // GetHTTPReq returns the http request object that can be used to send a event with payload to function service
 func GetHTTPReq(clientset kubernetes.Interface, funcName, namespace, eventNamespace, method, body string) (*http.Request, error) {
-	svc, err := clientset.CoreV1().Services(namespace).Get(funcName, metav1.GetOptions{})
+
+	funcPort, err := GetFunctionPort(clientset, namespace, funcName)
 	if err != nil {
-		return nil, fmt.Errorf("Unable to find the service for function %s", funcName)
+		return nil, err
 	}
-	funcPort := strconv.Itoa(int(svc.Spec.Ports[0].Port))
+
 	req, err := http.NewRequest(method, fmt.Sprintf("http://%s.%s.svc.cluster.local:%s", funcName, namespace, funcPort), strings.NewReader(body))
 	if err != nil {
 		return nil, fmt.Errorf("Unable to create request %v", err)
